feat: extract some logic and make reusable
