+++ src/quick-lane/quick-lane.service.ts
@@ -1,16 +1,23 @@
 import { ApolloQueryResult } from 'apollo-boost';
 import { get } from 'lodash-es';
 
-import { AssignmentContent, AssignmentContentLabel } from '@viaa/avo2-types/types/assignment';
+import { AssignmentContentLabel } from '@viaa/avo2-types/types/assignment';
 import { CollectionSchema } from '@viaa/avo2-types/types/collection';
 import { ItemSchema } from '@viaa/avo2-types/types/item';
-import { UserProfile, UserSchema } from '@viaa/avo2-types/types/user';
 
 import { ItemsService } from '../admin/items/items.service';
 import { AssignmentLayout } from '../assignment/assignment.types';
 import { CollectionService } from '../collection/collection.service';
 import { CustomError } from '../shared/helpers';
+import { quickLaneUrlRecordToObject } from '../shared/helpers/quick-lane-url-record-to-object';
 import { dataService } from '../shared/services';
+import {
+	QuickLaneInsertResponse,
+	QuickLaneQueryResponse,
+	QuickLaneUpdateResponse,
+	QuickLaneUrlObject,
+	QuickLaneUrlRecord,
+} from '../shared/types';
 
 import {
 	GET_QUICK_LANE_BY_CONTENT_AND_OWNER,
@@ -19,69 +26,7 @@ import {
 	UPDATE_QUICK_LANE,
 } from './quick-lane.gql';
 
-// Typing
-
-export interface QuickLaneUrl {
-	id: string;
-	title: string;
-	content?: AssignmentContent;
-	content_id?: string;
-	content_label?: AssignmentContentLabel;
-	owner?: QuickLaneUrlOwner;
-	owner_profile_id?: string;
-	created_at?: string;
-	updated_at?: string;
-}
-
-export interface QuickLaneUrlOwner extends Pick<UserProfile, 'id' | 'avatar'> {
-	usersByuserId: Pick<UserSchema, 'full_name'>;
-}
-
-export interface QuickLaneUrlObject extends QuickLaneUrl {
-	view_mode: AssignmentLayout;
-}
-
-export interface QuickLaneUrlRecord extends QuickLaneUrl {
-	view_mode: 'full' | 'without_description';
-}
-
-export interface QuickLaneQueryResponse {
-	app_quick_lanes: QuickLaneUrlRecord[];
-}
-
-export interface QuickLaneInsertResponse {
-	insert_app_quick_lanes: QuickLaneMutateResponse;
-}
-
-export interface QuickLaneUpdateResponse {
-	update_app_quick_lanes: QuickLaneMutateResponse;
-}
-
-export interface QuickLaneMutateResponse {
-	affected_rows: number;
-	returning: QuickLaneUrlRecord[];
-}
-
 // Mappers
-
-const quickLaneUrlRecordToObject = (record: QuickLaneUrlRecord) => {
-	const mapped = ({ ...record } as unknown) as QuickLaneUrlObject;
-
-	switch (record.view_mode) {
-		case 'full':
-			mapped.view_mode = AssignmentLayout.PlayerAndText;
-			break;
-
-		case 'without_description':
-			mapped.view_mode = AssignmentLayout.OnlyPlayer;
-			break;
-
-		default:
-			break;
-	}
-
-	return mapped;
-};
 
 const quickLaneUrlObjectToRecord = (object: QuickLaneUrlObject) => {
 	const mapped = ({ ...object } as unknown) as QuickLaneUrlRecord;
@@ -227,7 +172,7 @@ export class QuickLaneService {
 		}
 	}
 
-	static async fetchQuickLaneByContentAndOwnerId(
+	static async fetchQuickLanesByContentAndOwnerId(
 		contentId: string,
 		contentLabel: AssignmentContentLabel,
 		profileId: string
