+++ src/shared/helpers/generate-quick-lane-href.ts
@@ -1,0 +1,7 @@
+import { generatePath } from 'react-router';
+
+import { APP_PATH } from '../../constants';
+
+export const generateQuickLaneHref = (id: string): string => {
+	return generatePath(APP_PATH.QUICK_LANE.route, { id });
+};
